pagar-prestamo
Carga de sesion y pago de prestamo

+++ Core/Entities/Loan.cs
@@ -18,8 +18,8 @@ namespace Core.Entities
         public decimal CapitalToShow { get; set; }
         [Required(ErrorMessage = "Este campo es Obligatorio")]
         [DisplayFormat(DataFormatString = "{0:n0}")]
-        public decimal Quote { get; set; }
-        public decimal RemainingPayments { get; set; }
+        public int Quote { get; set; }
+        public int RemainingPayments { get; set; }
         public string Notes { get; set; }
         [Required(ErrorMessage = "Este campo es Obligatorio")]
         public int ClientId { get; set; }
